make camera more configurable and try to implement some more shapes
The extra shapes are in various states of "not working" at the moment.

Box works the best, but needs to be positioned to exist at the offset, while the cylinder rendering seeems very broken, even though the effect of the cylinder existing somehow shows on the ground sphere.

The rounded box looks like a regular box, but a lil weird.

Will work on these more later. For now planning to continue with "the next week"