an incomplete implementation of aabb and bvh
The BVH implementation is particularly challenging to translate from
C++ to Rust for no real reason other than its hard to figure out what
its doing, what its supposed to be doing, and what it should be in Rust.

The concepts are fairly simple: Build a bounding box around the spheres,
Those boxes are defined by ranges. Extend the ranges to combine them
into larger and larger boxes. Make that a tree structure and use it.

I think a petgraph-based approach would be interesting, but it could
also be easily possible using a tree structure with Box and such.

Overall, we aren't having performance issues anyway, and translating the
C++ implementation for this just isn't that fun or interesting, so I'm
going to drop it for now.